third_party/WebKit/LayoutTests/fast/css/style-and-stylesheet-important.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
    #p { color: red !important; }
</style>
<p id="p" style="color: green !important;">This text should be green.</p>
<script>
description("Style attribute !important wins over stylesheet !important");

shouldBeEqualToString("getComputedStyle(p).color", "rgb(0, 128, 0)");
</script>
